--- a/src/joint_constraints.rs
+++ b/src/joint_constraints.rs
@@ -1,7 +1,7 @@
 use crate::bindings::Bindings;
 use crate::patterns::FORM_SEPARATOR;
 use std::cmp::Ordering;
-use crate::constraints::VarConstraints;
+use crate::constraints::{VarConstraint, VarConstraints};
 use crate::umiaq_char::UmiaqChar;
 
 /// Compact representation of the relation between (sum) and (target).
@@ -232,82 +232,87 @@ pub(crate) fn parse_joint_constraints(equation: &str) -> Option<JointConstraints
 /// eliminates huge amounts of search, especially for long chains of unconstrained vars.
 pub fn propagate_joint_to_var_bounds(vcs: &mut VarConstraints, jcs: &JointConstraints) {
     for jc in &jcs.as_vec {
-        // Only handle equality constraints; others (<=, >=, !=) don’t propagate as cleanly
         if jc.rel != RelMask::EQ { continue; }
 
-        // Gather current bounds
+        // Cache per-var (min,max) and aggregate sums
         let mut sum_min = 0usize;
-        let mut sum_max: Option<usize> = Some(0);
+        let mut sum_max_opt: Option<usize> = Some(0);
+
         let mut mins: Vec<(char, usize)> = Vec::with_capacity(jc.vars.len());
-        let mut maxs: Vec<(char, Option<usize>)> = Vec::with_capacity(jc.vars.len());
+        let mut maxs: Vec<(char, usize)> = Vec::with_capacity(jc.vars.len());
 
         for &v in &jc.vars {
-            let vc = vcs.get(v);
-            let li = vc.and_then(|c| c.min_length).unwrap_or(1);
-            let ui = vc.and_then(|c| c.max_length);
+            let (li, ui) = vcs.bounds(v);
             sum_min += li;
-            sum_max = match (sum_max, ui) {
-                (Some(a), Some(b)) => Some(a + b),
-                _ => None, // if any ui is unbounded, the total is unbounded
-            };
+
+            // Track finite sum of maxes; if any is ∞, the group max is unbounded.
+            if ui == VarConstraint::DEFAULT_MAX {
+                sum_max_opt = None;
+            } else {
+                sum_max_opt = sum_max_opt.map(|a| a + ui);
+            }
+
             mins.push((v, li));
             maxs.push((v, ui));
         }
 
-        // Case 1: minimums already force exact solution → fix every var to its min
+        // Case 1: exact by mins
         if sum_min == jc.target {
             for (v, li) in mins {
-                vcs.ensure(v).set_exact_len(li);
+                vcs.ensure_entry_mut(v).set_exact_len(li);
             }
             continue;
         }
 
-        // Case 2: maximums already force exact solution → fix every var to its max
-        if let Some(total_max) = sum_max {
-            if total_max == jc.target {
+        // Case 2: exact by finite maxes
+        if let Some(sum_max) = sum_max_opt {
+            if sum_max == jc.target {
                 for (v, ui) in maxs {
-                    if let Some(ui) = ui {
-                        vcs.ensure(v).set_exact_len(ui);
-                    }
+                    vcs.ensure_entry_mut(v).set_exact_len(ui);
                 }
                 continue;
             }
         }
 
-        // Case 3: generic tightening for each variable individually
+        // Case 3: generic tightening
         for &v in &jc.vars {
-            let li = vcs.get(v).and_then(|c| c.min_length).unwrap_or(1);
-            let ui = vcs.get(v).and_then(|c| c.max_length);
+            let (li, ui) = vcs.bounds(v);
 
-            // Compute bounds for "other" variables
-            let sum_other_min: usize = jc.vars.iter()
+            // Σ other mins
+            let sum_other_min: usize = jc.vars
+                .iter()
                 .filter(|&&w| w != v)
-                .map(|&w| vcs.get(w).and_then(|c| c.min_length).unwrap_or(1))
+                .map(|&w| vcs.bounds(w).0)
                 .sum();
 
-            let mut sum_other_max: Option<usize> = Some(0);
+            // Σ other finite maxes (None if any is ∞)
+            let mut sum_other_max_opt: Option<usize> = Some(0);
             for &w in jc.vars.iter().filter(|&&w| w != v) {
-                let w_ui = vcs.get(w).and_then(|c| c.max_length);
-                sum_other_max = match (sum_other_max, w_ui) {
-                    (Some(a), Some(b)) => Some(a + b),
-                    _ => None, // unbounded if any is unbounded
-                };
+                let (_, w_ui) = vcs.bounds(w);
+                if w_ui == VarConstraint::DEFAULT_MAX {
+                    sum_other_max_opt = None;
+                    break;
+                } else {
+                    sum_other_max_opt = sum_other_max_opt.map(|a| a + w_ui);
+                }
             }
 
-            // Joint constraint gives tighter lower/upper bounds
-            let lower_from_joint = match sum_other_max {
+            let lower_from_joint = match sum_other_max_opt {
                 Some(s) => jc.target.saturating_sub(s),
-                None => 0, // unbounded above for others ⇒ no new lower bound
+                None => 0, // others can stretch arbitrarily
             };
             let upper_from_joint = jc.target.saturating_sub(sum_other_min);
 
-            let e = vcs.ensure(v);
-            e.min_length = Some(std::cmp::max(li, lower_from_joint));
-            e.max_length = Some(std::cmp::min(ui.unwrap_or(usize::MAX), upper_from_joint));
+            // Tighten and store
+            let new_min = li.max(lower_from_joint);
+            let new_max = ui.min(upper_from_joint);
+
+            let e = vcs.ensure_entry_mut(v);
+            e.min_length = Some(new_min);
+            e.max_length = Some(new_max);
         }
     }
 }
-
 
 #[cfg(test)]
 mod tests {