… src/scan_hints.rs