We create individual variable constraints from joint constraints, where applicable. This can help optimize some queries.

Example: `cargo run --release -- 'ABCDEFGHIJKLMN;|ABCDEFGHIJKLMN|=14;!=ABCDEFGHIJKLMN' -n 10`

Before: 
```
Loaded 113096 words in 0.050s; solved in 10.030s (2 tuples).
```

After:
```
Loaded 113096 words in 0.063s; solved in 0.016s (2 tuples).
```

I also took this opportunity to simplify the code somewhat and ditch a bunch of TODOs

I do think this will mess with some of the other PRs because the major change here is that bounds are just `usize` as opposed to `Option<usize>`